requirements.txt
alembic==1.7.3
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dotenv==0.19.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.4.25
Werkzeug==2.0.2
zipp==3.6.0
python-decouple==3.5